% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r2.R
\name{add_r2_from_vcftools}
\alias{add_r2_from_vcftools}
\title{Add r2 values to GRanges object}
\usage{
add_r2_from_vcftools(geno_ld, gr_snps, pvalue = NULL, status = NULL)
}
\arguments{
\item{geno_ld}{The file *.geno.ld produced by `vcftools` or a
\code{data.frame} from the file loaded with read.table.}

\item{gr_snps}{A \code{GRanges} object.}

\item{pvalue}{The name of the column to use as pvalue. Default: NULL (the
first numerical metadata column will be used)}

\item{status}{Name of the column for the SNP status (Genotype or Imputed).
  If value is \code{NULL}, all the SNPs will be used to find the best SNP,
  otherwise only SNPs with \code{Genotyped} value will be used.
  Default: NULL.}
}
\value{
A \code{numeric} vector corresponding to the pairwise LD of best SNP
  versus all the other SNPs.
}
\description{
This function will:
  1) Find the best SNPs in \code{gr_snps} using the \code{pvalue} parameter.
  2) Will extract relevant lines from \code{geno_ld}.
  3) Create \code{r2} vector.
}

