\name{manhattan-package}
\alias{manhattan-package}
\alias{manhattan}
\docType{package}
\title{
\packageTitle{manhattan}
}
\description{
\packageDescription{manhattan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{manhattan}
\packageIndices{manhattan}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{manhattan}

Maintainer: \packageMaintainer{manhattan}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
