% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Make a Manhattan plot}
\usage{
manhattan(gwas, build = c("hg18", "hg19", "hg38"), color1 = "black",
  color2 = "grey")
}
\arguments{
\item{gwas}{A GWAS study. Must have chrom, pos, and y columns}

\item{build}{Genomic build. Currently supports hg18, hg19, and hg38}

\item{color1}{Color for odd-numbered chromosomes}

\item{color2}{Color for even-numbered chromosomes}
}
\value{
a ggplot object that makes a Manhattan plot
}
\description{
Make a Manhattan plot
}
\details{
\code{manhattan} is a wrapper around \code{ggplot}. It uses a few tricks to transform a genomic axis to a scatterplot axis. For instance, chr2:1 would be the length of chromosome 1 plus 1, chr3:1 would be chromosome 1 plus chromosome 2 plus 1, so on and so forth. It is important to specify the genomic build (e.g. hg19) so that `manhattan` can make the correct transformation. It positions the chromosome labels on the x-axis according to these transformations.
}
